requirements.txt
PySide2>=5.15.2
opencv-python>=4.5.0
pytesseract>=0.3.10
easyocr>=1.6.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
Pillow>=9.0.0
openpyxl>=3.0.0
numpy>=1.21.0
python-Levenshtein>=0.12.0
ultralytics>=8.0.0  # YOLO v8 for improved table detection (optional)
